requirements/base.txt
flake8==3.8.3
isort==5.4.2
pre-commit==2.7.1
python-dotenv==0.17.0
elasticsearch==7.12.0
jupyter==1.0.0